get Owner func back into routes.go
